jsinspector: Vendor base64 encoding, remove folly::base64 dependency
Summary:
Roll our own base64 encoding and revert D59685218, which fixed the missing Folly dependency implied by D54309633.

I assumed Folly base64 was already elsewhere in RN but given it isn't, and we only need simple, non-perf-sensitive encoding for the debugger (not the SIMD or delegated implementations, or decoding), it might be best to just include our own encoder.

Reviewed By: cipolleschi

Differential Revision: D58323859